fix: update basic, renderVideo, and primitives examples to support webgpu.
