+++ example/renderVideo.js
@@ -24,6 +24,7 @@ import { LoaderElement } from './utils/LoaderElement.js';
 const ENV_URL = 'https://raw.githubusercontent.com/gkjohnson/3d-demo-data/master/hdri/phalzer_forest_01_1k.hdr';
 const MODEL_URL = 'https://raw.githubusercontent.com/gkjohnson/3d-demo-data/main/models/bao-robot/bao-robot.glb';
 const CREDITS = 'Model by DailyArt on Sketchfab';
+const PREVIEW_RENDER_DELAY = 100;
 
 // CCapture seems to replace the requestAnimationFrame callback which breaks the ability to render and
 // use CanvasCapture.
@@ -70,6 +71,7 @@ const params = {
 		} );
 
 		// reinitialize recording variables
+		pathTracer.renderDelay = 0;
 		recordedFrames = 0;
 		regenerateScene();
 		rebuildGUI();
@@ -78,6 +80,8 @@ const params = {
 	stop: () => {
 
 		CanvasCapture.stopRecord();
+		pathTracer.renderDelay = PREVIEW_RENDER_DELAY;
+		pathTracer.reset();
 		recordedFrames = 0;
 		rebuildGUI();
 
@@ -107,7 +111,7 @@ async function init() {
 	pathTracer = new WebGPUPathTracer( renderer );
 	pathTracer.filterGlossyFactor = 0.25;
 	pathTracer.tiles.set( params.tiles, params.tiles );
-	pathTracer.renderDelay = 0;
+	pathTracer.renderDelay = PREVIEW_RENDER_DELAY;
 	pathTracer.minSamples = 1;
 	pathTracer.fadeDuration = 0;
 
@@ -269,6 +273,8 @@ function animate() {
 
 		camera.updateMatrixWorld();
 
+		pathTracer.renderSample();
+
 		// if we're recording and we hit the target samples then record the frame step the animation forward
 		if ( isRecording && pathTracer.samples >= params.samples ) {
 
@@ -279,6 +285,7 @@ function animate() {
 			if ( recordedFrames >= params.frameRate * params.duration ) {
 
 				CanvasCapture.stopRecord();
+				pathTracer.renderDelay = PREVIEW_RENDER_DELAY;
 
 				recordedFrames = 0;
 				rebuildGUI();
@@ -297,8 +304,6 @@ function animate() {
 			regenerateScene();
 
 		}
-
-		pathTracer.renderSample();
 
 	}
